requirements.txt
numpy
python-can
uptime
PyQT5
mypy
black
isort
cantools
